requirements.txt
# LifeKB Backend Dependencies
# Purpose: Python package dependencies for the privacy-first journaling API

# Core API Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database & Auth
supabase==2.0.0
asyncpg==0.29.0
sqlalchemy==2.0.23

# AI & Embeddings
openai==1.3.5
numpy==1.24.3

# HTTP & Async
httpx==0.25.2
aiofiles==23.2.1

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.0.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2 